+++ packages/config/src/theme/AccordionTitleText.ts
@@ -1,0 +1,6 @@
+import { createStyle } from '@gluestack-style/react';
+
+export const AccordionTitleText = createStyle({
+  flex: 1,
+  textAlign: 'left',
+});