+++ packages/config/src/theme/Actionsheet.ts
@@ -3,4 +3,7 @@ import { createStyle } from '@gluestack-style/react';
 export const Actionsheet = createStyle({
   width: '$full',
   height: '$full',
+  _web: {
+    pointerEvents: 'none',
+  },
 });