+++ packages/config/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @gluestack-ui/config
 
+## 1.0.6
+
+### Patch Changes
+
+- fix: tooltip pointer events
+
+## 1.0.5
+
+### Patch Changes
+
+- fix: version bump @gluestack-style/react
+- Updated dependencies
+  - @gluestack-ui/themed@1.0.20
+
+## 1.0.4
+
+### Patch Changes
+
+- fix: toast
+- Updated dependencies
+  - @gluestack-ui/themed@1.0.18
+
 ## 1.0.3
 
 ### Patch Changes